Change `numLocalBytes` to `u32` (#500)
When we have 8192 `i64` locals, the required bytes are 65536.

+++ source/m3_function.h
@@ -83,7 +83,7 @@ typedef struct M3Function
     u16                     numRetAndArgSlots;
 
     u16                     numLocals;                              // not including args
-    u16                     numLocalBytes;
+    u32                     numLocalBytes;
 
     bool                    ownsWasmCode;
